requirements.txt
django
djangorestframework
pydantic
pydantic_ai
python-dotenv